+++ ReferenceApiWeblessUmbraco/Areas/HelpPage/App_Start/HelpPageConfig.cs
@@ -3,15 +3,12 @@
 ////#define Handle_PageResultOfT
 
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Net.Http.Headers;
-using System.Reflection;
-using System.Web;
 using System.Web.Http;
+
 #if Handle_PageResultOfT
 using System.Web.Http.OData;
 #endif
@@ -19,9 +16,10 @@ using System.Web.Http.OData;
 namespace ReferenceApiWeblessUmbraco.Areas.HelpPage
 {
     /// <summary>
-    /// Use this class to customize the Help Page.
-    /// For example you can set a custom <see cref="System.Web.Http.Description.IDocumentationProvider"/> to supply the documentation
-    /// or you can provide the samples for the requests/responses.
+    ///     Use this class to customize the Help Page.
+    ///     For example you can set a custom <see cref="System.Web.Http.Description.IDocumentationProvider" /> to supply the
+    ///     documentation
+    ///     or you can provide the samples for the requests/responses.
     /// </summary>
     public static class HelpPageConfig
     {