+++ UmbracoVault/TypeHandlers/MemberTypeHandler.cs
@@ -4,7 +4,6 @@ using System.Linq;
 
 using Umbraco.Core;
 using Umbraco.Core.Models;
-using Umbraco.Web;
 
 using UmbracoVault.Exceptions;
 using UmbracoVault.Extensions;
@@ -12,7 +11,7 @@ using UmbracoVault.Extensions;
 namespace UmbracoVault.TypeHandlers
 {
     /// <summary>
-    /// Responsible for converting Member types
+    ///     Responsible for converting Member types
     /// </summary>
     public class MemberTypeHandler : ITypeHandler
     {
@@ -21,17 +20,23 @@ namespace UmbracoVault.TypeHandlers
             var result = typeof(T).CreateWithNoParams<T>();
 
             if (result == null)
+            {
                 throw new ConstructorUnavailableException(typeof(T));
+            }
 
             var member = GetMember(input.ToString());
             if (member != null)
+            {
                 Vault.Context.FillClassProperties(result, (alias, recursive) =>
                 {
                     if (!member.HasProperty(alias))
+                    {
                         return null;
+                    }
 
                     return member.GetValue(alias);
                 });
+            }
 
             return result;
         }