+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Extended the current UmbracoContext implementation to include a secondary implem
  
  * Added `PropertyInfo` parameter to `FillClassProperties` to support additional behavior for libraries that extend Vault.  [#25]
 
+## v1.2.2
+ * Fixes an issue where models implementing an interface would not properly have their values set [#29]
+ * Fixes an issue where explicit interface implementations would not properly have their values set [#29]
+>>>>>>> master
+
 ## v1.2.1
 
  * Fixes an issue where classes that used type inheritance would cause duplicate results to return in calls to `GetUmbracoEntityAliasesFromType` [#12]